Add Base Resource Consumption
The base now consumes resources, with a progress bar showing
how much of the resource is left to be consumed before another
resource will be consumed. If all resources are consumed, the base no
longer provides light, and the player can no longer regen sanity.
If the player provides resources to the base, then the light will
return, and sanity will regen.

Signed-off-by: Alex Schendel <alex.schendel@intel.com>