Remove executables from tracked files
Executables take up a lot of space on the repo, so it is best to keep
them separate from the source code. (Makes clones/pulls much longer)
I added the Linux and Windows executables to the .gitignore so they
will not be tracked by git. Instead, I have uploaded them under the
releases tab as version 1.0.0

Signed-off-by: Alex Schendel <alex.schendel@intel.com>